Move components AppCookieBanner, AppCookieCustomization to sub path cookies
